#5 added temporary implementation of favicon
The goal of this commit is to establish an icon for this page.
- Icon must react to light/dark theme.
- It also serves as a logo on main page, so it must also react to dark mode toggle

Icon on main page was imported very "hacky". Unfortunatly, when imported with <img/> svg didn't react to theme change from toggle. The SVG must be "declared" inside the page, so that `applyPreferredColorScheme` function from `src/util` could do its job. Thus, currently icon on mainpage is imported with `dangerouslySetInnerHTML`. Is there a better way to do it?