move github link to footer
It feels a bit out of place in header: no use like site functionality or call to action. Thus it was moved to "Powered by".
The Logo component was moved up a level in hierarchy, now it is used both in header and footer. Didn't feel right to make a whole new folder for it, since it is styled by header and footer individually. It is still usefull as it helps to encapsulate logic of light/dark img links and context usage for this selection.